Move Myspace seed content into feature data
